EOD check-in

+++ app/sqs_to_bq.py
@@ -2,11 +2,13 @@ import json
 import multiprocessing
 import pprint
 
-from lib import schema
 import boto3
 from absl import app
 from absl import flags
 from absl import logging
+from google.cloud import bigquery
+from lib import schema
+#from lib import bq_utils
 
 # Define args
 FLAGS = flags.FLAGS
@@ -17,12 +19,17 @@ flags.mark_flag_as_required('queue_url')
 
 # Global vars
 sqs = boto3.resource('sqs')
+bq_client = bigquery.Client()
 
 
 def stream_rows(json_rows: str, table_name: str, dataset_name: str):
-  print(json_rows)
+  test = json.loads(json_rows)
+  print(len(test))
+  #print(json_rows)
   print(table_name)
   print(dataset_name)
+
+  return True
 
 
 def main(argv):
@@ -49,10 +56,11 @@ def main(argv):
       dataset_name = message.message_attributes.get('dataset', {}).get(
           'StringValue')
 
-      stream_rows(message.body, table_name, dataset_name)
-      message.delete()
-      available_messages = int(
-          sqs_queue.attributes.get('ApproximateNumberOfMessages'))
+      streaming_status = stream_rows(message.body, table_name, dataset_name)
+      if streaming_status:
+        message.delete()
+        available_messages = int(
+            sqs_queue.attributes.get('ApproximateNumberOfMessages'))
 
   logging.info('Queue processing complete!')
